Split out PathComponent struct
This commit moves the `PathComponent` struct into its own file to
unclutter the mod.rs file.
